third_party/blink/web_tests/http/tests/worklet/dynamic-import-expected.txt

CONSOLE MESSAGE: This test logs a result once from each PaintWorkletGlobalScope
CONSOLE ERROR: TypeError: import() is disallowed on WorkletGlobalScope.
CONSOLE ERROR: TypeError: import() is disallowed on WorkletGlobalScope.
This is a testharness.js-based test.
Harness: the test ran to completion.
